Revert "Add basic TAP tests for the low-level backup method"
This reverts commit 99b4a63bef94.  The test is proving to be unstable,
so revert it for now.

One of the failures seen involves the cluster started without the
backup_label, where the archives of the primary are overwritten, causing
recovery failures on Windows.  This is simple to fix, but there is
another issue that's creeping behind in the form of an "invalid data in
file" ERROR when parsing the backup_label for the second recovery case,
as an effect of the backup_label data written after retrieving its
contents from pg_backup_stop().
_
Per buildfarm member sidewinder.

(cherry picked from commit 6cb1b632b354518a595f1b23bfbac28a798f7f1c)